fix: massive fix of models
