+++ src/interfaces/files.interface.ts
@@ -1,0 +1,14 @@
+import { Document } from 'mongoose';
+
+interface File extends Document {
+    filename: string;
+    contentType: string;
+    length: number;
+    chunkSize: number;
+    uploadDate: Date;
+    aliases: string[];
+    metadata: any;
+    md5: string;
+}
+
+export default File;
