+++ src/interfaces/historics.interface.ts
@@ -1,0 +1,10 @@
+import { Document } from 'mongoose';
+import mongoose from "mongoose";
+
+interface Historics extends Document {
+    id: number;
+    idUser: mongoose.Types.ObjectId;
+    idVideo: mongoose.Types.ObjectId;
+}
+
+export default Historics;
