+++ src/interfaces/roles.interface.ts
@@ -1,0 +1,12 @@
+
+import { Document } from 'mongoose';
+import mongoose from "mongoose";
+
+interface Role extends Document {
+    id: number;
+    name: string;
+    permission: string;
+    idUsers: mongoose.Types.ObjectId;
+}
+
+export default Role;
