+++ src/models/users.models.ts
@@ -1,18 +1,5 @@
-import mongoose, { Schema, Document } from 'mongoose';
-
-interface User extends Document {
-  id: number;
-  firstName: string;
-  lastName: string;
-  email: string;
-  dateOfBirth: Date;
-  username: string;
-  password: string;
-  createdAt: Date;
-  status: boolean;
-  language: string;
-  profileImage: {type: Schema.Types.ObjectId, ref:'Files'};
-}
+import mongoose, { Schema } from 'mongoose';
+import User from '../interfaces/users.interface';
 
 const UserSchema = new Schema<User>({
   id: {
